Prevented crash on unanticipated name changes
DBusServiceRegistry now can cope with reversed order of NameOwnerChanged
signals. Also, as a side effect, now it copes well with name transfers.
Tests for both were added, though the test for the first issue requires
one private member of the ServiceRegistry to be public and therefore is
commented.
On the road one small performance improvement and some formatting issues
were tackled.