Use MainLoopContext name to lookup bus type
Without this fix it is not possible to use the system bus when running in a
single thread with a custom main loop.

MainLoopContext::getName() is passed as connection id to DBusConnection()
so it makes sense to use it to lookup bus type in DBusAddressTranslator.
(Connection id is used to lookup bus type when using a separate dispatch
thread.)

Signed-off-by: Martin Ejdestig <martin.ejdestig@volvocars.com>