Add files via upload
Max number of replies that can be pending for each connection count is limited to 127 and 50000 in case of SYSTEM bus  and SESSION bus respectively. Currently when an api call of type fireAndForget is made, it's DBUS_HEADER_FLAG_NO_REPLY_EXPECTED flag is false so the count for pending call is incremented for every call of type fireAndForget and since no reply is sent for such calls this count will keep on increasing till it reaches the configuration specified limit. Once the limit is reached, no more calls can be made. Since fireAndForget api calls don't expect any reply/response we can set the flag DBUS_HEADER_FLAG_NO_REPLY_EXPECTED to FALSE.