Extend DBusObjectManager to use GetAll
The DBusObjectManager is specified by freedesktop to return as each
value in inner dict the same dict that would be returned by the
org.freedesktop.DBus.Properties.GetAll() method for that combination of
object path and interface. Only if an interface has no properties, an
empty dict will be returned.